# docs/phase-7/openapi.yaml
openapi: 3.1.0
info:
  title: InkEcho API
  version: 1.0.0
  description: |
    REST API for InkEcho multiplayer party game.
    Primary UI uses Server Actions — this spec covers Route Handlers.
    Realtime events via Ably — see docs/phase-0/10-realtime-events.md
  contact:
    name: InkEcho
servers:
  - url: https://inkecho.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ink_player_session
    authCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    cronBearer:
      type: http
      scheme: bearer
  parameters:
...
    RoomCode:
      name: code
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Z0-9]{6}$'
...
  schemas:
...
    RoomSettings:
      type: object
      properties:
        maxPlayers:
          type: integer
          minimum: 3
          maximum: 12
          default: 8
        minPlayers:
          type: integer
          default: 3
        roundCount:
          type: integer
          minimum: 1
          maximum: 3
          default: 1
        describeTimerSec:
          type: integer
          minimum: 30
          maximum: 120
          default: 60
        drawTimerSec:
          type: integer
          minimum: 60
          maximum: 180
          default: 90
        profanityFilter:
          type: boolean
          default: false
        allowSpectators:
          type: boolean
          default: true
    Participant:
      type: object
      required: [playerId, displayName, role, isReady, connectionStatus]
      properties:
        playerId:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          type: string
          enum: [HOST, PLAYER, SPECTATOR]
        isReady:
          type: boolean
        connectionStatus:
          type: string
          enum: [ONLINE, RECONNECTING, OFFLINE]
        avatarUrl:
          type: string
          nullable: true
    RoomSnapshot:
      type: object
      required: [id, code, status, visibility, hostPlayerId, settings, participants]
      properties:
        id:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [LOBBY, IN_PROGRESS, REVEAL, CLOSED]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        hostPlayerId:
          type: string
        settings:
          $ref: '#/components/schemas/RoomSettings'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        spectatorCount:
          type: integer
...
paths:
...
  /rooms/{code}:
    get:
      tags: [Rooms]
      summary: Get room snapshot
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RoomSnapshot'
        '404':
          description: Not found
    patch:
      tags: [Rooms]
      summary: Update room settings (host)
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                settings:
                  $ref: '#/components/schemas/RoomSettings'
      responses:
        '200':
          description: OK
        '403':
          description: Not host
    delete:
      tags: [Rooms]
      summary: Close room (host)
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '204':
          description: Closed
...